mudar tipo de recompensa: de Double para BigDecimal

+++ src/main/java/br/com/infnet/guilda_dos_aventureiros/dto/relatorios/RankingProjection.java
@@ -2,10 +2,12 @@ package br.com.infnet.guilda_dos_aventureiros.dto.relatorios;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 
+import java.math.BigDecimal;
+
 @JsonPropertyOrder({"nomeAventureiro", "totalParticipacoes", "totalRecompensas", "totalDestaques"})
 public interface RankingProjection {
     String getNomeAventureiro();
     Long getTotalParticipacoes();
-    Double getTotalRecompensas();
+    BigDecimal getTotalRecompensas();
     Long getTotalDestaques();
 }
